flask/requirement.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
priority==2.0.0
psycopg2==2.9.6
PyJWT==2.10.1
python-dotenv==1.0.0
SQLAlchemy==2.0.17
typing_extensions==4.12.2
Werkzeug==2.3.6
wsproto==1.2.0
